Variable scope and understanding of order of events was wrong
1) I thought that by declaring var inside a top level function that I
would then be able to modify that variable inside nested functions,
but it doesn't work like this, so need to make the variable global
outside of all functions

2) I also didn't understand the order of the events. I assumed because
the job lot was wrapped in a jQuery function that it would do the
getFeed bit and after that was complete it would then do the following
function where it appends the result. But this isn't the case as the
getFeed doesn't finish before the append occurs. Therefore use the
ajaxComplete function to wait until it has.