Basically make some CSS changes and improvements
Meant to sort the max width out ages ago because the paragraphs were too wide on the
blog posts with a wide browser window. But the problem I had is that I didn't want a
max width on the archive page as the more fluid that is the better (for the
floating layout).

So had to change the templating a bit to support different styles on the archive
page and the rest of the pages. Which meant spiltting up the header includes into
three. This is actually better because means I only include the javascript for the
Archive page on the Archive page instead of including it everywhere unnecessarily.

Also, expand use of webfonts and use everywhere. Why not hey?